renaming and cleaning a bunch of files
