+++ src/index.html
@@ -312,17 +312,15 @@
                             id="theme-settings"
                             class="flat-btn settings-switch-btn-cont"
                             effect="theme-switch">
-                        <data-display id="theme-btn"
-                                      class="settings-switch-bnt"
-                                      prop="theme">Dark</data-display>
+                        <theme-value id="theme-btn"
+                                     class="settings-switch-bnt">Dark</theme-value>
                     </button>
                     <button is="effect-button"
                             id="msystem-settings"
                             class="flat-btn settings-switch-btn-cont"
                             effect="measurement-switch">
-                        <data-display id="measurement-btn"
-                                      class="settings-switch-bnt"
-                                      prop="measurement">Metric</data-display>
+                        <measurement-value id="measurement-btn"
+                                           class="settings-switch-bnt">Metric</measurement-value>
                     </button>
                 </div> <!-- end settings btn row -->
