+++ src/views/ant-device-scan.js
